Fix race condition in instance.update sample test
The test case test_create_delete_server_with_instance_update failed randomly
because the last expected instance.update notification was not received due
to a race condition.

The test polls the REST API and waits for the instance to reach ACTIVE state.
However the last instance.update is sent from the instance.save after the
state of the instance is updated in the DB. Therefore it is possible that
the test saw ACTIVE instance state before the last instance.update
notification is emitted.

This patch introduce an extra wait call to receive the last notification
as well.

Closes-Bug: #1657087
Change-Id: I438bbb6de0733d0c3fd8361fdc4149c88abb53aa