libvirt: Mock is_job_complete in test_driver
Various tests were modelling the return of blockJobInfo rather than
mocking guest.BlockDevice.is_job_complete. While I'm generally very
much in favour of this kind of functional testing, distributing it
throughout individual tests is not maintainable. If we're going to do
this we need to maintain a minimal simulation library, which we don't
have.

This change updates all tests which use the return of blockJobInfo to
implicitly affect the return of is_job_complete. This allows us to
fix is_job_complete without breaking a large number of tests.

Change-Id: I676d28759da55ce678572fee135d55468dbe6a41