Allow placement endpoint interface to be set
This patch exposes the "interface" option for ks_filter to allow
the placement API to be connected on a specific endpoint interface.

The previous was to force "public", which is default for keystoneauth.
The default for the placement service mirrors this value.

Change-Id: Ic996e596f8473c0b8626e8d0e92e1bf58044b4f8
(cherry picked from commit 1257d338db480ba237f6878cc478d072b28f7695)