Explain how allow_resize_to_same_host is useful
With the ServerGroupAffinityFilter enabled, you need to also set
allow_resize_to_same_host to True. Mentioning this in the
description for that config setting.

ServerGroupAffinity is a scheduler setting that places all
servers with the group on the same host. (It is the opposite
of AntiAffinity that PREVENTS them from being on the same
host.) Consequently, if you try to resize a server that is
in an Affinity group (ie, it is affined) it will fail as the
current rules IMMEDIATELY remove the current server from
the list of possible scheduling targets (unless you have enabled
"allow_resize_to_same_host".) Since none of the other servers
can be candidates because of the Affinity rule, the resize
fails entirely.

Change-Id: I77a90ffdfe465d99ee8596375430ee960048d566
Closes-Bug: #1660090