doc: update the man page entry for nova-manage db sync
The docs for this command were woefully out of date. This
change clarifies the wording since we're really upgrading
the schema in the database, not "synchronizing" it even
though that's the name of the command - the migration scripts
run the 'upgrade' method only. Also, it was saying this was
how you created the main database, which is totally wrong
since this command doesn't create the database. The deployer
creates the database, this command migrates the schema on that
database.

Then I also added the arguments along with a description of each
and noted that this will by default upgrade the cell0 schema.

Change-Id: I3d40cb32c325bc3f665e727b921082efba043192