Clarify the [cells] config option help
People are easily confused into thinking that for cells v2
they need to configure things in the [cells] group in
nova.conf, but the cells options are strictly for cells v1
functionality, which we don't want people using anymore.

This change adds some wording to those options to make it
clear that they are for cells v1 and that is not a
recommended way to deploy Nova.

Change-Id: Ib1f0caf956d0bbb76d2e5a8e6d5ce7de12e240b2