neutron: retrieve physical network name from a multi-provider network
Multi-provider networks are a combination of vxlan and vlan networks.
such a network has a special "segments" field that holds a list
of networks. Each element of this list has the same structure as the 'provider
network attributes'.
These attributes are:
- provider:network_type
- provider:physical_network
- provider:segmentation_id

This patch will retrieve the relevant physical network name from a
multi-provider network

Change-Id: Icb9a546e6fcbf399fcef3e9ea686b35ec3817cd5
Closes-Bug: #1659467