libvirt: fix and break up _test_attach_detach_interface
The detach_interface flow in this test was broken because
it wasn't mocking out domain.detachDeviceFlags so the xml
it was expecting to be passed to that method wasn't actually
being verified. The same thing is broken in test
test_detach_interface_device_with_same_mac_address because
it copies the other broken test code.

This change breaks apart the monster attach/detach test method
and converts the detach_interface portion to mock and fixes
the broken assertion.

test_detach_interface_device_with_same_mac_address is just
fixed, not converted to mock.

Change-Id: I6d9a975876c5652ef544c587f65b1bdd1543848b
Related-Bug: #1607714