Bump minimum required libvirt/qemu versions for Pike
Based on the libvirt distro support matrix wiki [1] this change
bumps the minimum required version of libvirt to 1.2.9 and
QEMU to 2.1.0. These were both advertised as the next minimums
since Newton, we just never made the change in Ocata.

The next minimum libvirt version is set to 1.3.1 and the next
minimum QEMU version is set to 2.5.0, which is what we gate
on with Ubuntu 16.04 but also falls within the distro support
matrix for a representative set of other supported distros.

[1] https://wiki.openstack.org/wiki/LibvirtDistroSupportMatrix

Change-Id: I9a972e3fde2e4e552f6fc98350820c07873c3de3