api-ref: fix instance action 'message' description
There was a typo in the description of this parameter
and it also didn't read very clearly, so I've just
re-written the description of this.

On a side note, it's a bit odd that this is a required
parameter in the response (Tempest requires it in the
response schema) even though it's only set if the
action failed.

Change-Id: Ic84031119e2ffb6b1fb0b0a193967b2823464651