Make server_external_events cells-aware
This fixes the external events api to be aware of instances in cells.
Without this, we can't deliver the event from neutron to the compute
host and thus the instance never unpauses on startup. With this patch,
I can create an instance in a cell other than the one that nova-api
has as its default.

Related to blueprint cells-aware-api
Change-Id: I2077e539d30764ac629c2b1418887f123aa02194