Add helper method to add additional data about policy rule.
We need operators to understand what the policy rules mean without
looking at the code. To fix this we improve the description set in the
oslo.policy RuleDefault.

This patch introduces a helper method create_rule_default to ensure we
create consistent descriptions for all the policy rules. This helper
method would be changed to use DocumentedRuleDefault once new version
of oslo.policy is released.

Also, this patch adds description to policies in
nova/policies/server_tags.py

blueprint policy-docs

Change-Id: Ibf403bb5c05f3cef93f5a7cbb9eab95381a733a8