Fix s390 "connector not found" issue
This patch introduces a try-catch block on a call to the os-brick
library which may throw a ValueError. os_brick will throw it when
a driver is not supported, so ignoring it and skipping its addition
to the driver list should be the expected behavior. This fix is
required for s390, where nova is currently broken without it.

The issue was resolved in Ocata with changes including [1] and [2],
as well as a change to os_brick [3]. The backport to newton [4] was
abandoned as it requires a raise in the os_brick minimum version.
This fix is a revision of [5], which is essentially a lightweight
version of [1] that only addresses the issue at hand.

[1]: https://review.openstack.org/#/c/395067/
[2]: https://review.openstack.org/#/c/394979/
[3]: https://review.openstack.org/#/c/375415/
[4]: https://review.openstack.org/#/c/428386/
[5]: https://review.openstack.org/#/c/394425/

Change-Id: I638f12ce3deab0338a8d2652398c35acb9280f72
Closes-Bug: 1639239