libvirt: drop MIN_LIBVIRT_HUGEPAGE_VERSION
The minimum required version of libvirt (1.2.9) covers this case, so we
can drop the compatibility code.

This change effectively means it is not possible to have hugepages if
you don't have a NUMA topology, and the '_has_hugepage_support' function
is now essentially a less-strict version of 'has_numa_support' and can
be removed. Do this, and remove the corresponding tests.

Change-Id: I7ceb82293700d331c2422a15439e70d09a4dc96a