re-orphan flavor after rpc deserialization
The flavor objects stored in the instance object should not be used
to lazy load any data from the db as they are not connected to real
flavors in the db. These flavor objects are orphaned when created
but durin rpc deserialization every deserialized nova object
automatically gets a valid context. This can lead to unintended
lazy loading of the projects field of such flavor objects during
notification payload generation.

This patch orphans the flavor objects stored in the instance object
after every deserialization.

Change-Id: I458f81931bad7874e951e1c0fd464d149f61b244
Related-Bug: #1653221