Check for 204 case in DynamicVendorData
Before b9587587c2f3d6f871dec07562c16e1d82bc05e6 we were doing
an explicit check on the response status codes, but after
that change, we just rely on the bool override in requests.Reponse
which is True for any repsonse code less than 400. In the 204 case
we won't have response text, so we should handle that before trying
to load the json repsonse text. Because TypeError and ValueError are
handled it would just result in a warning if you got a 204, unless
CONF.api.vendordata_dynamic_failure_fatal was True in which case
it's a failure (as shown in the unit test update here).

Change-Id: Iba95fd0112ae4510ef21fdb44fcb281b14567c07
Closes-Bug: #1669084