Add functional test for ip filtering with regex
There was a bug in Tempest where it would filter the
list of servers by fixed IP address and expected to
only get back one server but sometimes got more than
one. That was because the IP of one server was
a prefix of the IP of another server, and because of
the regex used in the _ip_filter method in the compute
API it will return all matches.

This is how the IP filtering has always worked, but
it's not clear we had a solid test coverage of this
scenario so a functional test was written to confirm
the expected behavior and so that we don't regress it
since it's not documented in the api-ref (yet).

Change-Id: If0e6d7ea7ac61a8480bf7f0288b263902536dc59
Related-Bug: #1668828