fdatasync() downloaded images before use
Image download closes the filehandle of the downloaded image before
use, but doesn't fdatasync it. This means that in the event of a host
crash we may be left with only the file metadata when the host comes
back up: i.e. an empty file of the correct size. Nova cannot detect
this data corruption, so operator intervention is required.

By adding an fdatasync() before closing the file handle we ensure that
the downloaded file is either entirely present, or entirely not
present.

See also change I33bd99b0, which fixes this issue for downloads
requiring a subsequent conversion step.

Partial-Bug: #1669844
Change-Id: Id9905a87f16f66530623800e33e2581c555ae81d