Make os-availability-zones know about cells
This removes the direct-to-object listing of Service objects by the
availability zone routines. Instead, it implements a cell-scanning
approach in HostAPI and uses that for consistent behavior.

The modified behavior of service_get_all() in HostAPI was originally
in another patch, but is moved here because of dependency sequencing.
The HostAPI patch later in this series uses this as well.

Related to blueprint cells-aware-api
Change-Id: I65c2b436bef8837c0e10a5e502e9cd598d9aa0c3