Remove unused placement_database config options
The placement_database config options were added in Newton
but the actual code to use the options was reverted and is
not used. The help text actually has a typo (Ocata was 15.0.0,
not 14.0.0) and, more importantly, the help text makes it sound
like you should really be setting up a separate database for the
placement service but we don't actually use these options, we use
the api_database options for all of the placement data models.

To avoid confusion until this is actually supported, let's just
remove the options which should have been cleaned up as part of
39fb302fd9c8fc57d3e4bea1c60a02ad5067163f anyway.

Conflicts:
        nova/tests/functional/api/openstack/placement/fixtures.py

NOTE(mriedem): The conflict is due to not having change
a598af9de6536d8b66046fa329d9960edc706aed in stable/ocata.

Change-Id: I31293ac4689630e4113588ab2c6373cf572b8f38
Closes-Bug: #1670419
(cherry picked from commit 7439e9473617a871c032107ab85b1b623369975f)