Update docstrings for legacy notification methods
It's not particularly clear which methods send legacy
unversioned notifications since the docstrings are not
clear about that, nor is the method name.

I came across this while trying to figure out what
puts the "compute." prefix on the legacy unversioned
notifications and whether or not that was on the
versioned notifications (it's not).

This just updates the docstrings to be more clear.

Change-Id: Ic85f4f9d31a719368495fc0dcd07ace043816693