Add regression test for bug 1670627
This adds a functional regression test for bug 1670627.

This is the recreate scenario. Patches that are proposed to
fix the bug will build on top of this and change it's assertions
to know when it's properly fixed.

Change-Id: I872a3fd5cfd3dd869f74cd3fd0aa5da411b1fec3
Related-Bug: #1670627
(cherry picked from commit 500282d867a15b441068d9327ddcb9ae69f41b7d)