conf: Deprecated 'defer_iptables_apply'
This is *mostly* used for nova-network and therefore can be deprecated.
There is one corner case where we could conceivably use this with
neutron and the libvirt driver, but it's only ever going to work with
with Linux bridge, is probably broken and should just be removed in
favour of neutron variants of this functionality. A note is included
detailing this, just in case people do want to spend the time getting to
the bottom of this, but I wouldn't recommend that.

The help text for this option is improved based on information I found
while researching the option and mostly taken from the commit message
for commit 8f1c54ce.

Change-Id: I33607453b3174192a33d9d56e203227bc9237f31