Avoid lazy-loading projects during flavor notification
This makes the flavor notification not load the flavor's projects
if they are not already set. Even though we attempted to orphan the
instance-bound flavors in a recent change[1], there are other times
where we were loading projects purely for the notification. This
lazy-load physically will not work if we're in the cell, and is
never something we want to do when we're an instance-bound flavor.

This makes us check to see if projects are loaded, and if so, clone
the flavor to notify on, and set projects on that clone to an
empty list.

[1]: This effectively reverts I458f81931bad7874e951e1c0fd464d149f61b244
which solved the problem in some places, but not all.

Change-Id: I0f80f101a528ab9cda81d025cf316f2ac16bdcb7