Introduce fast8 tox target
The pep8 target has the little known "-HEAD" argument
which makes it only run flake8 on the changes since the
last change in the history.

Even though that's nice, it's sometimes hard to remember
the -HEAD argument. Plus pep8 is running some other
checks on things like the docs.

So if all you want is the pep8 -- -HEAD behavior, this
is a simplied convenient form of it.

Change-Id: I97b277d8ac95ca3ef8379570e0e3886df26263e1