Add a functional regression/recreate test for bug 1671648
This adds a test which recreates the regression bug introduced
in Ocata where build retries are not populated when creating
instances in conductor for cells v2.

The change that fixes the bug will go on top of this and modify
the test to show the bug is fixed.

Change-Id: Ie9e955d79b4e1441092183135b3f70b003c94db5
Related-Bug: #1671648
(cherry picked from commit 72e1506101b131b51fbe77acc0af19f36899c28d)