Add a note about force_hosts only ever having a single value
The fact force_hosts and force_nodes are lists in the
scheduler filter properties and RequestSpec object is really
confusing when you actually look at what populates those lists,
which is a single value from the API when specifying an
availability zone during server create:

  nova boot .. --availability-zone ZONE:HOST:NODE

You can only have at most one host or node to force the
server build. It appears the confusion was originally
introduced way back in Essex with this change:

16ea348a1623f055809d0d9b7fe9f046515b5dd1

Maybe that was for future proofing, or maybe it was
somehow related to pre-Essex features, or simply a typo,
we don't really know. But what's clear is you can't
provide a list of hosts to force the server build onto
via the compute REST API, at least not with vanilla
unmodified Nova.

This change just adds some comments to the RequestSpec
object so that future generations trying to sort this
out don't have to stumble along as much as I did.

Change-Id: I5d17fb04de22f680fe9a6cbc1e06e946abfc5aff