Count server groups to check quota
This changes server groups from a ReservableResource to a
CountableResource and replaces quota reserve/commit/rollback with
check_deltas accordingly.

A new configuration option [quota]/recheck_quota has also been added
to control whether quota should be checked a second time after
resources have been created, to prevent allowing quota to be exceeded
as a result of racing requests.

Co-Authored-By: melanie witt <melwittt@gmail.com>

Part of blueprint cells-count-resources-to-check-quota-in-api

Change-Id: If87c84001673e5e463136435327044cc06f87a17