Add get_count_by_vm_state() to InstanceList object
This adds a method for getting a count of instances in a specified
vm_state. We want to use this to limit the number of pending resizes
per user for checking quota. When a user reverts a resize to a smaller
flavor, there is a possibility of exceeding cores and ram quota
because quota is not checked during a revert, so as not to block
reverts.

Co-Authored-By: melanie witt <melwittt@gmail.com>

Part of blueprint cells-count-resources-to-check-quota-in-api

Change-Id: I7ef402b8fed000d01cb4c7d49f7ecf48d8c19e55