Add check_deltas() and limit_check_project_and_user() to Quotas
This adds two new methods to the Quotas object in preparation for the
switch to counting resources for checking quota.

The existing limit_check() method only supports checking quota against
per-user limits because currently, the only CountableResources that
exist are not ever counted across a project.

To implement resource counting for all resources, we will change each
resource to a countable resource and some of the resources will be
counted across a project and across a user and checked against
per-project and per-user limits, respectively. To do this, we add a new
method: count_as_dict that returns a count as a dictionary to support
multiple counts (over a project and over a user) and a new method:
limit_check_project_and_user() that supports checking quota against
both per-project and per-user limits.

The check_deltas() method is a wrapper around the count_as_dict() and
limit_check_project_and_user() calls where only the deltas are passed
in and the counting and limit check are done inside the wrapper. This
cleans up the call sites and enables us to potentially make the
count + check atomic in the future behind the check_deltas() interface.

A change has also been made to the _validate_method_call_check_resource
methods to let them take the Resources dict as an argument instead of
assuming and hard-coding the global QuotaEngine's resources. This was
mainly needed for unit testing because in order to cover both the
per-project and per-user limit checking, a CountableResource counted
across both a project and a user is needed, and currently none exist
in the global QuotaEngine. Resources will be changed over in
subsequent patches.

Co-Authored-By: melanie witt <melwittt@gmail.com>

Part of blueprint cells-count-resources-to-check-quota-in-api

Change-Id: Ib52a21836c4638cb9f7a5a55a5befbf0963fe6bf