remove hacking rule that enforces log translation
All of the po files for log translation have been deleted from the
Nova tree as of I3206d8bcbffe7ef1d84acf0fac6b423a26fccf94 which has
made all of this infrastructure unused. The mailing list discussions
are pretty much all coming back to the fact that no one ever really
wanted this feature, and it was pushed in purely as a product
requirement checklist.

This removes hacking enforcement of this now unused feature, and will
allow people to start deleting the i18n markup without coming afoul of
hacking.

Change-Id: I9c334162fe1799e7b24563fdc11256b91bbafc9f