Log a warning if there is only one cell when listing instances
We expect to have at least two cells when listing instances, cell0
and the main cell. If the operator deployed cell0 and started nova-api
before creating the main cell we need to log a warning informing them
that something is wrong and they will need to restart nova-api to
refresh the cache once the main cell is created.

Change-Id: I8fcfe0554e0f263224afd0d39403463df80d4cce
Related-Bug: #1669473