libvirt: mark some Image backend methods as abstract
Change I6c68693be818219f78d6fcda019b867066396b29 shows us
that we should be marking Image methods as abstract if they
actually are required to be implemented by a subclass. Since
the Image class itself is an ABC, let's take advantage and
mark some of the other required methods in the base class
as abstract methods so we don't hit this same issue again with
a new backend.

Change-Id: Id5129e48d1ec3b23b552b066916a968c02953988
Related-Bug: #1670642