Remove the stevedore extension point for server create
This patch remove the stevedore extension point for server
create.

This patch move all the extension points into ServersController
class attribute 'server_create_func_list'. This is for
backward-compatible with the stevedore extension interface.

The final goal is that merging all of those extended code into
the ServerController. So this is a middle step for refactor.

Partial-implement-blueprint api-no-more-extensions-pike

Change-Id: I9a8f56037b35e89543626922221b910ab8b1917e