nova-manage: Deprecate 'project', 'account' commands
Most of these operations can be accomplished using the REST API [1],
with the 'quota_usage_refresh' command being the sole exception. This
particular command is duplicated in a new category - quota - thus
allowing us to fully deprecate the entire 'project' and 'account'
categories.

Note that the '(project|account) quota_usage_refresh' command has been
named 'quota refresh' and not 'quota usage_refresh'. This is because
cliff, which will be used by 'nova-manage' in the future, does not
support underscores in command names.

[1] https://developer.openstack.org/api-ref/compute/#quota-sets-os-quota-sets

Change-Id: Ic02a96802d2a38ed9c4e00991a6bd2088dfcffc6