Removes potentially bad exit value from accepted list in Quobyte volume driver
This disallows exit code 4 in the list of valid exit codes of the execute
call for mounting Quobyte volumes. As exit code 0 is the default value
for the allowed exit codes the whole check_exit_codes argument is
omitted.
Furthermore this updates the (dis-)connect volume sync locks to be
Quobyte specific.

Closes-Bug: #1679976

Change-Id: I87b74535bd1a2045948a56d1648cd587146facb7