Short-circuit local delete path for cells v2 and InstanceNotFound
When we're going down the local delete path for cells v2 in the API
and instance.destroy() fails with an InstanceNotFound error, we are
racing with a concurrent delete request and know that the instance
is alread deleted, so we can just return rather than fall through to
the rest of the code in the _delete() method, like for BDMs and
console tokens.

Conflicts:
      nova/compute/api.py

NOTE(mriedem): The conflict is due to not having change
edf51119fa59ff8a3337abb9107a06fa33d3c68f in stable/ocata.

Change-Id: I58690a25044d2804573451983323dde05be9e5d6
Closes-Bug: #1680211
(cherry picked from commit 5a9cc2fb7af3e3a9db44646bbd23cfcfb16891f5)