Change livesnapshot to true by default
Because of past qemu bugs we set this to default off and required a
workaround to enable it. However these issues appear to be fixed, and
we should enable this by default for people (with the ability to turn
it off if it's causing crashes in their environment). Metacloud/Cisco
has been running default this way the entire time without issue.

Change-Id: I4545cc99683d311c41c8d6d8dc9787ea6ec6cbc9