Handle new hosts for updating instance info in scheduler
As of change 791cf0643401f72ce834e580938057e325945169 we have
to go through host and cell mappings to get to a cell database
for host (compute node) and instance information.

When a new compute service starts up it casts to update_instance_info
and sends an empty list, which triggers the scheduler to try and
lookup all instances on that host. If the new compute host is not
yet mapped in a cell, we'll get a HostMappingNotFound error.

We can handle this case in both the HostManager and ComputeManager
by not casting from the compute if there is no information to send,
and we can also handle it in the HostManager (for older computes)
but just dealing with the HostMappingNotFound gracefully.

This information eventually self-heals via normal operation.

Change-Id: I7cec2eff35c0615534fcb4d5148f75721824172e
Closes-Bug: #1680661